+++ bindings/internal/raw/libraries/endpointsecurity/bridge/endpointsecurity_bridge.m
@@ -142,10 +142,10 @@ void * endpointsecurity_fn_es_exec_fd(void * event, uint32_t index, void **outEx
 }
 
 // ID: objc-sym EndpointSecurity.es_subscribe
-int64_t endpointsecurity_fn_es_subscribe(void * client, void * events, uint32_t event_count, void **outException) {
+int32_t endpointsecurity_fn_es_subscribe(void * client, void * events, uint32_t event_count, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_subscribe(client, events, event_count);
+            int32_t _result = (int32_t)es_subscribe(client, events, event_count);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -155,10 +155,10 @@ int64_t endpointsecurity_fn_es_subscribe(void * client, void * events, uint32_t 
 }
 
 // ID: objc-sym EndpointSecurity.es_unsubscribe
-int64_t endpointsecurity_fn_es_unsubscribe(void * client, void * events, uint32_t event_count, void **outException) {
+int32_t endpointsecurity_fn_es_unsubscribe(void * client, void * events, uint32_t event_count, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_unsubscribe(client, events, event_count);
+            int32_t _result = (int32_t)es_unsubscribe(client, events, event_count);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -168,10 +168,10 @@ int64_t endpointsecurity_fn_es_unsubscribe(void * client, void * events, uint32_
 }
 
 // ID: objc-sym EndpointSecurity.es_unsubscribe_all
-int64_t endpointsecurity_fn_es_unsubscribe_all(void * client, void **outException) {
+int32_t endpointsecurity_fn_es_unsubscribe_all(void * client, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_unsubscribe_all(client);
+            int32_t _result = (int32_t)es_unsubscribe_all(client);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -181,10 +181,10 @@ int64_t endpointsecurity_fn_es_unsubscribe_all(void * client, void **outExceptio
 }
 
 // ID: objc-sym EndpointSecurity.es_subscriptions
-int64_t endpointsecurity_fn_es_subscriptions(void * client, void * count, void * subscriptions, void **outException) {
+int32_t endpointsecurity_fn_es_subscriptions(void * client, void * count, void * subscriptions, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_subscriptions(client, count, subscriptions);
+            int32_t _result = (int32_t)es_subscriptions(client, count, subscriptions);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -194,10 +194,10 @@ int64_t endpointsecurity_fn_es_subscriptions(void * client, void * count, void *
 }
 
 // ID: objc-sym EndpointSecurity.es_respond_auth_result
-int64_t endpointsecurity_fn_es_respond_auth_result(void * client, void * message, int64_t result, bool cache, void **outException) {
+int32_t endpointsecurity_fn_es_respond_auth_result(void * client, void * message, int32_t result, bool cache, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_respond_auth_result(client, message, result, cache);
+            int32_t _result = (int32_t)es_respond_auth_result(client, message, result, cache);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -207,10 +207,10 @@ int64_t endpointsecurity_fn_es_respond_auth_result(void * client, void * message
 }
 
 // ID: objc-sym EndpointSecurity.es_respond_flags_result
-int64_t endpointsecurity_fn_es_respond_flags_result(void * client, void * message, uint32_t authorized_flags, bool cache, void **outException) {
+int32_t endpointsecurity_fn_es_respond_flags_result(void * client, void * message, uint32_t authorized_flags, bool cache, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_respond_flags_result(client, message, authorized_flags, cache);
+            int32_t _result = (int32_t)es_respond_flags_result(client, message, authorized_flags, cache);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -220,10 +220,10 @@ int64_t endpointsecurity_fn_es_respond_flags_result(void * client, void * messag
 }
 
 // ID: objc-sym EndpointSecurity.es_mute_process
-int64_t endpointsecurity_fn_es_mute_process(void * client, void * audit_token, void **outException) {
+int32_t endpointsecurity_fn_es_mute_process(void * client, void * audit_token, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_mute_process(client, audit_token);
+            int32_t _result = (int32_t)es_mute_process(client, audit_token);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -233,10 +233,10 @@ int64_t endpointsecurity_fn_es_mute_process(void * client, void * audit_token, v
 }
 
 // ID: objc-sym EndpointSecurity.es_mute_process_events
-int64_t endpointsecurity_fn_es_mute_process_events(void * client, void * audit_token, void * events, uint64_t event_count, void **outException) {
+int32_t endpointsecurity_fn_es_mute_process_events(void * client, void * audit_token, void * events, uint64_t event_count, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_mute_process_events(client, audit_token, events, event_count);
+            int32_t _result = (int32_t)es_mute_process_events(client, audit_token, events, event_count);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -246,10 +246,10 @@ int64_t endpointsecurity_fn_es_mute_process_events(void * client, void * audit_t
 }
 
 // ID: objc-sym EndpointSecurity.es_unmute_process
-int64_t endpointsecurity_fn_es_unmute_process(void * client, void * audit_token, void **outException) {
+int32_t endpointsecurity_fn_es_unmute_process(void * client, void * audit_token, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_unmute_process(client, audit_token);
+            int32_t _result = (int32_t)es_unmute_process(client, audit_token);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -259,10 +259,10 @@ int64_t endpointsecurity_fn_es_unmute_process(void * client, void * audit_token,
 }
 
 // ID: objc-sym EndpointSecurity.es_unmute_process_events
-int64_t endpointsecurity_fn_es_unmute_process_events(void * client, void * audit_token, void * events, uint64_t event_count, void **outException) {
+int32_t endpointsecurity_fn_es_unmute_process_events(void * client, void * audit_token, void * events, uint64_t event_count, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_unmute_process_events(client, audit_token, events, event_count);
+            int32_t _result = (int32_t)es_unmute_process_events(client, audit_token, events, event_count);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -272,10 +272,10 @@ int64_t endpointsecurity_fn_es_unmute_process_events(void * client, void * audit
 }
 
 // ID: objc-sym EndpointSecurity.es_muted_processes
-int64_t endpointsecurity_fn_es_muted_processes(void * client, void * count, void * audit_tokens, void **outException) {
+int32_t endpointsecurity_fn_es_muted_processes(void * client, void * count, void * audit_tokens, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_muted_processes(client, count, audit_tokens);
+            int32_t _result = (int32_t)es_muted_processes(client, count, audit_tokens);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -285,10 +285,10 @@ int64_t endpointsecurity_fn_es_muted_processes(void * client, void * count, void
 }
 
 // ID: objc-sym EndpointSecurity.es_muted_processes_events
-int64_t endpointsecurity_fn_es_muted_processes_events(void * client, void * muted_processes, void **outException) {
+int32_t endpointsecurity_fn_es_muted_processes_events(void * client, void * muted_processes, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_muted_processes_events(client, muted_processes);
+            int32_t _result = (int32_t)es_muted_processes_events(client, muted_processes);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -309,10 +309,10 @@ void endpointsecurity_fn_es_release_muted_processes(void * muted_processes, void
 }
 
 // ID: objc-sym EndpointSecurity.es_mute_path
-int64_t endpointsecurity_fn_es_mute_path(void * client, const char * path, int64_t type_, void **outException) {
+int32_t endpointsecurity_fn_es_mute_path(void * client, const char * path, int32_t type_, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_mute_path(client, path, type_);
+            int32_t _result = (int32_t)es_mute_path(client, path, type_);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -322,10 +322,10 @@ int64_t endpointsecurity_fn_es_mute_path(void * client, const char * path, int64
 }
 
 // ID: objc-sym EndpointSecurity.es_mute_path_events
-int64_t endpointsecurity_fn_es_mute_path_events(void * client, const char * path, int64_t type_, void * events, uint64_t event_count, void **outException) {
+int32_t endpointsecurity_fn_es_mute_path_events(void * client, const char * path, int32_t type_, void * events, uint64_t event_count, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_mute_path_events(client, path, type_, events, event_count);
+            int32_t _result = (int32_t)es_mute_path_events(client, path, type_, events, event_count);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -335,10 +335,10 @@ int64_t endpointsecurity_fn_es_mute_path_events(void * client, const char * path
 }
 
 // ID: objc-sym EndpointSecurity.es_mute_path_prefix
-int64_t endpointsecurity_fn_es_mute_path_prefix(void * client, const char * path_prefix, void **outException) {
+int32_t endpointsecurity_fn_es_mute_path_prefix(void * client, const char * path_prefix, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_mute_path_prefix(client, path_prefix);
+            int32_t _result = (int32_t)es_mute_path_prefix(client, path_prefix);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -348,10 +348,10 @@ int64_t endpointsecurity_fn_es_mute_path_prefix(void * client, const char * path
 }
 
 // ID: objc-sym EndpointSecurity.es_mute_path_literal
-int64_t endpointsecurity_fn_es_mute_path_literal(void * client, const char * path_literal, void **outException) {
+int32_t endpointsecurity_fn_es_mute_path_literal(void * client, const char * path_literal, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_mute_path_literal(client, path_literal);
+            int32_t _result = (int32_t)es_mute_path_literal(client, path_literal);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -361,10 +361,10 @@ int64_t endpointsecurity_fn_es_mute_path_literal(void * client, const char * pat
 }
 
 // ID: objc-sym EndpointSecurity.es_unmute_all_paths
-int64_t endpointsecurity_fn_es_unmute_all_paths(void * client, void **outException) {
+int32_t endpointsecurity_fn_es_unmute_all_paths(void * client, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_unmute_all_paths(client);
+            int32_t _result = (int32_t)es_unmute_all_paths(client);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -374,10 +374,10 @@ int64_t endpointsecurity_fn_es_unmute_all_paths(void * client, void **outExcepti
 }
 
 // ID: objc-sym EndpointSecurity.es_unmute_all_target_paths
-int64_t endpointsecurity_fn_es_unmute_all_target_paths(void * client, void **outException) {
+int32_t endpointsecurity_fn_es_unmute_all_target_paths(void * client, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_unmute_all_target_paths(client);
+            int32_t _result = (int32_t)es_unmute_all_target_paths(client);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -387,10 +387,10 @@ int64_t endpointsecurity_fn_es_unmute_all_target_paths(void * client, void **out
 }
 
 // ID: objc-sym EndpointSecurity.es_unmute_path
-int64_t endpointsecurity_fn_es_unmute_path(void * client, const char * path, int64_t type_, void **outException) {
+int32_t endpointsecurity_fn_es_unmute_path(void * client, const char * path, int32_t type_, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_unmute_path(client, path, type_);
+            int32_t _result = (int32_t)es_unmute_path(client, path, type_);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -400,10 +400,10 @@ int64_t endpointsecurity_fn_es_unmute_path(void * client, const char * path, int
 }
 
 // ID: objc-sym EndpointSecurity.es_unmute_path_events
-int64_t endpointsecurity_fn_es_unmute_path_events(void * client, const char * path, int64_t type_, void * events, uint64_t event_count, void **outException) {
+int32_t endpointsecurity_fn_es_unmute_path_events(void * client, const char * path, int32_t type_, void * events, uint64_t event_count, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_unmute_path_events(client, path, type_, events, event_count);
+            int32_t _result = (int32_t)es_unmute_path_events(client, path, type_, events, event_count);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -413,10 +413,10 @@ int64_t endpointsecurity_fn_es_unmute_path_events(void * client, const char * pa
 }
 
 // ID: objc-sym EndpointSecurity.es_muted_paths_events
-int64_t endpointsecurity_fn_es_muted_paths_events(void * client, void * muted_paths, void **outException) {
+int32_t endpointsecurity_fn_es_muted_paths_events(void * client, void * muted_paths, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_muted_paths_events(client, muted_paths);
+            int32_t _result = (int32_t)es_muted_paths_events(client, muted_paths);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -437,10 +437,10 @@ void endpointsecurity_fn_es_release_muted_paths(void * muted_paths, void **outEx
 }
 
 // ID: objc-sym EndpointSecurity.es_invert_muting
-int64_t endpointsecurity_fn_es_invert_muting(void * client, int64_t mute_type, void **outException) {
+int32_t endpointsecurity_fn_es_invert_muting(void * client, int32_t mute_type, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_invert_muting(client, mute_type);
+            int32_t _result = (int32_t)es_invert_muting(client, mute_type);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -450,10 +450,10 @@ int64_t endpointsecurity_fn_es_invert_muting(void * client, int64_t mute_type, v
 }
 
 // ID: objc-sym EndpointSecurity.es_muting_inverted
-int64_t endpointsecurity_fn_es_muting_inverted(void * client, int64_t mute_type, void **outException) {
+int32_t endpointsecurity_fn_es_muting_inverted(void * client, int32_t mute_type, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_muting_inverted(client, mute_type);
+            int32_t _result = (int32_t)es_muting_inverted(client, mute_type);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -463,10 +463,10 @@ int64_t endpointsecurity_fn_es_muting_inverted(void * client, int64_t mute_type,
 }
 
 // ID: objc-sym EndpointSecurity.es_clear_cache
-int64_t endpointsecurity_fn_es_clear_cache(void * client, void **outException) {
+int32_t endpointsecurity_fn_es_clear_cache(void * client, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_clear_cache(client);
+            int32_t _result = (int32_t)es_clear_cache(client);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -476,10 +476,10 @@ int64_t endpointsecurity_fn_es_clear_cache(void * client, void **outException) {
 }
 
 // ID: objc-sym EndpointSecurity.es_new_client
-int64_t endpointsecurity_fn_es_new_client(void * client, void * handler, void **outException) {
+int32_t endpointsecurity_fn_es_new_client(void * client, void * handler, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_new_client(client, (es_handler_block_t)handler);
+            int32_t _result = (int32_t)es_new_client(client, (es_handler_block_t)handler);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
@@ -489,10 +489,10 @@ int64_t endpointsecurity_fn_es_new_client(void * client, void * handler, void **
 }
 
 // ID: objc-sym EndpointSecurity.es_delete_client
-int64_t endpointsecurity_fn_es_delete_client(void * client, void **outException) {
+int32_t endpointsecurity_fn_es_delete_client(void * client, void **outException) {
     @autoreleasepool {
         @try {
-            int64_t _result = (int64_t)es_delete_client(client);
+            int32_t _result = (int32_t)es_delete_client(client);
             return _result;
         } @catch (NSException *_ex) {
             if (outException) *outException = (__bridge void *)[_ex retain];
